refactor: pkg/client/proof/types to pkg/client/types
