+++ pkg/relayer/session/session.go
@@ -8,6 +8,7 @@ import (
 	cosmostypes "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/pokt-network/poktroll/pkg/client"
+	clienttypes "github.com/pokt-network/poktroll/pkg/client/types"
 	"github.com/pokt-network/poktroll/pkg/observable"
 	"github.com/pokt-network/poktroll/pkg/observable/channel"
 	"github.com/pokt-network/poktroll/pkg/observable/logging"
@@ -46,7 +47,7 @@ type relayerSessionsManager struct {
 	blockClient client.BlockClient
 
 	// supplierClients is used to create claims and submit proofs for sessions.
-	supplierClients *proof.SupplierClientMap
+	supplierClients *clienttypes.SupplierClientMap
 
 	// storesDirectory points to a path on disk where KVStore data files are created.
 	storesDirectory string
